+++ ChangeLog.md
@@ -10,6 +10,7 @@ Table:
 - Fix binary table property validation to be 64K bytes not 32K characters.
 - Does not error when table created is a substring of another table.
 - Correctly responds with status 404 on patch with non-existant entity.
+- Fix pagination when no rowkey in continuation token
 
 ## 2022.04 Version 3.17.1